+++ src/pages/404.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import Layout from '../components/Layout/Layout';
-import SEO from '../components/seo';
+import SEO from '../components/SEO';
 
 const NotFoundPage = () => (
   <Layout>
